requirements.txt
streamlit 
transformers 
torch 
sacrebleu 
sentencepiece 
panda